+++ server/tests/seed/readmodel/names.py
@@ -1,6 +1,6 @@
 from uuid import UUID
 
-from the_history_atlas.apps.domain.models.readmodel.tables import NameModel
+from the_history_atlas.apps.domain.models.history.tables import NameModel
 
 NAMES: list[NameModel] = [
     NameModel(
