+++ server/tests/seed/readmodel/people.py
@@ -1,7 +1,7 @@
 from typing import List
 from uuid import UUID
 
-from the_history_atlas.apps.domain.models.readmodel.tables import PersonModel
+from the_history_atlas.apps.domain.models.history.tables import PersonModel
 
 PEOPLE: List[PersonModel] = [
     PersonModel(
