+++ server/tests/seed/readmodel/places.py
@@ -1,7 +1,7 @@
 from typing import List
 from uuid import UUID
 
-from the_history_atlas.apps.domain.models.readmodel.tables import PlaceModel
+from the_history_atlas.apps.domain.models.history.tables import PlaceModel
 
 PLACES: List[PlaceModel] = [
     PlaceModel(
