+++ server/tests/seed/readmodel/tag_instances.py
@@ -1,6 +1,6 @@
 from uuid import UUID
 
-from the_history_atlas.apps.domain.models.readmodel.tables import TagInstanceModel
+from the_history_atlas.apps.domain.models.history.tables import TagInstanceModel
 
 TAG_INSTANCES: list[TagInstanceModel] = [
     TagInstanceModel(
