+++ server/the_history_atlas/apps/app_manager.py
@@ -1,14 +1,14 @@
 from the_history_atlas.apps.accounts.accounts import Accounts
 from the_history_atlas.apps.config import Config
 from the_history_atlas.apps.database import DatabaseApp
-from the_history_atlas.apps.readmodel import ReadModelApp
+from the_history_atlas.apps.history import HistoryApp
 
 
 class AppManager:
     config_app: Config
     database_app: DatabaseApp
     accounts_app: Accounts
-    readmodel_app: ReadModelApp
+    readmodel_app: HistoryApp
 
     def __init__(self, config_app: Config):
         self.config_app = config_app
@@ -17,7 +17,7 @@ class AppManager:
             config=self.config_app, database_client=self.database_app.client()
         )
 
-        self.readmodel_app = ReadModelApp(
+        self.readmodel_app = HistoryApp(
             config_app=self.config_app,
             database_client=self.database_app.client(),
         )
