+++ server/the_history_atlas/apps/domain/models/__init__.py
@@ -46,6 +46,6 @@ from the_history_atlas.apps.domain.models.nlp.text_analysis import (
     TextAnalysis,
     GetTextAnalysis,
 )
-from the_history_atlas.apps.domain.models.readmodel.queries.coords_by_name import (
+from the_history_atlas.apps.domain.models.history.queries.coords_by_name import (
     CoordsByName,
 )
