+++ server/the_history_atlas/apps/domain/models/history/__init__.py
@@ -1,0 +1,2 @@
+from the_history_atlas.apps.domain.models.history.default_entity import DefaultEntity
+from the_history_atlas.apps.domain.models.history.source import Source
