+++ server/the_history_atlas/apps/domain/models/history/default_entity.py
