+++ server/the_history_atlas/apps/domain/models/history/queries/__init__.py
@@ -1,0 +1,35 @@
+from the_history_atlas.apps.domain.models.history.queries.get_citation_by_id import (
+    GetCitationByID,
+    Citation,
+)
+from the_history_atlas.apps.domain.models.history.queries.get_summaries_by_id import (
+    GetSummariesByIDs,
+    Summary,
+)
+from the_history_atlas.apps.domain.models.history.queries.get_manifest import (
+    GetManifest,
+    Manifest,
+    Timeline,
+)
+from the_history_atlas.apps.domain.models.history.queries.get_entity_summaries_by_name import (
+    GetEntitySummariesByName,
+    EntitySummariesByName,
+)
+from the_history_atlas.apps.domain.models.history.queries.entity_summary import (
+    EntitySummary,
+)
+from the_history_atlas.apps.domain.models.history.queries.get_entity_ids_by_names import (
+    GetEntityIDsByNames,
+    EntityIDsByNames,
+)
+from the_history_atlas.apps.domain.models.history.queries.get_fuzzy_search_by_name import (
+    GetFuzzySearchByName,
+    FuzzySearchByName,
+)
+from the_history_atlas.apps.domain.models.history.queries.get_entity_summaries_by_ids import (
+    GetEntitySummariesByIDs,
+)
+from the_history_atlas.apps.domain.models.history.queries.get_place_by_coords import (
+    GetPlaceByCoords,
+    PlaceByCoords,
+)
