+++ server/the_history_atlas/apps/domain/models/history/queries/coords_by_name.py
