+++ server/the_history_atlas/apps/domain/models/history/queries/entity_summary.py
