+++ server/the_history_atlas/apps/domain/models/history/queries/get_entity_summaries_by_name.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from the_history_atlas.apps.domain.models.base_model import ConfiguredBaseModel
-from the_history_atlas.apps.domain.models.readmodel.queries.entity_summary import (
+from the_history_atlas.apps.domain.models.history.queries.entity_summary import (
     EntitySummary,
 )
