+++ server/the_history_atlas/apps/domain/models/history/queries/get_place_by_coords.py
