+++ server/the_history_atlas/apps/domain/models/history/tables/__init__.py
@@ -1,0 +1,13 @@
+from the_history_atlas.apps.domain.models.history.tables.source import SourceModel
+from the_history_atlas.apps.domain.models.history.tables.citation import CitationModel
+from the_history_atlas.apps.domain.models.history.tables.place import PlaceModel
+from the_history_atlas.apps.domain.models.history.tables.person import PersonModel
+from the_history_atlas.apps.domain.models.history.tables.time import TimeModel
+from the_history_atlas.apps.domain.models.history.tables.summary import SummaryModel
+from the_history_atlas.apps.domain.models.history.tables.tag_instance import (
+    TagInstanceModel,
+)
+from the_history_atlas.apps.domain.models.history.tables.tag_name_assoc import (
+    TagNameAssocModel,
+)
+from the_history_atlas.apps.domain.models.history.tables.name import NameModel
