+++ server/the_history_atlas/apps/domain/models/history/tables/summary.py
