--- server/the_history_atlas/apps/domain/models/readmodel/__init__.py
@@ -1,2 +1,0 @@
-from the_history_atlas.apps.domain.models.readmodel.default_entity import DefaultEntity
-from the_history_atlas.apps.domain.models.readmodel.source import Source
