--- server/the_history_atlas/apps/domain/models/readmodel/tables/__init__.py
@@ -1,13 +1,0 @@
-from the_history_atlas.apps.domain.models.readmodel.tables.source import SourceModel
-from the_history_atlas.apps.domain.models.readmodel.tables.citation import CitationModel
-from the_history_atlas.apps.domain.models.readmodel.tables.place import PlaceModel
-from the_history_atlas.apps.domain.models.readmodel.tables.person import PersonModel
-from the_history_atlas.apps.domain.models.readmodel.tables.time import TimeModel
-from the_history_atlas.apps.domain.models.readmodel.tables.summary import SummaryModel
-from the_history_atlas.apps.domain.models.readmodel.tables.tag_instance import (
-    TagInstanceModel,
-)
-from the_history_atlas.apps.domain.models.readmodel.tables.tag_name_assoc import (
-    TagNameAssocModel,
-)
-from the_history_atlas.apps.domain.models.readmodel.tables.name import NameModel
