+++ server/the_history_atlas/apps/history/__init__.py
@@ -1,0 +1,1 @@
+from the_history_atlas.apps.history.history_app import HistoryApp
