+++ server/the_history_atlas/apps/history/event_handler.py
@@ -29,13 +29,13 @@ from the_history_atlas.apps.domain.models import (
 )
 from the_history_atlas.apps.domain.models.events.meta_tagged import MetaTagged
 from the_history_atlas.apps.domain.types import Event
-from the_history_atlas.apps.readmodel.database import Database
-from the_history_atlas.apps.readmodel.errors import (
+from the_history_atlas.apps.history.database import Database
+from the_history_atlas.apps.history.errors import (
     UnknownEventError,
     DuplicateEventError,
     TagExistsError,
 )
-from the_history_atlas.apps.readmodel.trie import Trie
+from the_history_atlas.apps.history.trie import Trie
 
 log = logging.getLogger(__name__)
