--- server/the_history_atlas/apps/readmodel/__init__.py
@@ -1,1 +1,0 @@
-from the_history_atlas.apps.readmodel.read_model import ReadModelApp
